Set up basic duels configuration
