+++ src/main/java/io/github/neopixel/wrapper/games/duels/HypixelDuels.java
@@ -5,49 +5,11 @@ import io.github.neopixel.wrapper.games.HypixelGame;
 import io.github.neopixel.wrapper.util.JSONHandler;
 import java.util.Optional;
 
-public class HypixelDuels extends HypixelGame implements HypixelLootChestGame {
+public class HypixelDuels {
 
+    private final JSONHandler jsonHandler;
     protected HypixelDuels(JSONHandler jsonHandler) {
-        super(jsonHandler);
+        this.jsonHandler = jsonHandler;
     }
 
-    @Override
-    public String getGameID() {
-        return "DUELS";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Duels";
-    }
-
-    @Override
-    public boolean isRemoved() {
-        return false;
-    }
-
-    @Override
-    public final int getOpenedChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public final int getOpenedCommonChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public final int getOpenedRareChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public final int getOpenedEpicChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public final int getOpenedLegendaryChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
 }
